feat: refactor executor connectivity service for improved port handling and logging
